requirements.txt
python-telegram-bot==20.7
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.0.2
python-dotenv==1.0.1
